[acl] Add * fallback behavior for JQ selectors in contract reference (#129)
## Describe your changes and provide context
This adds the conservative fallback behavior for the JQ selectors when
processing a contract reference. Because we are in a contract reference,
we don't necessarily know that the JSON body is proper, therefore if a
JQ selector doesn't have a match, we have to conservatively assume that
it is because the JSON incorrectly doesn't include that field instead of
the dependency not being necessary. As a result, these JQ selectors will
modify the identifier template to be "*" (aka the conservative value for
that resource) in the case that the selector doesn't have a match

## Testing performed to validate your change
Updated unit tests to verify this behavior and also to test that we
properly discard JQ selectors when NOT in a contract reference.