[go.mod][lru] bump LRU to v2.0.1 for generics (#154)
## Describe your changes and provide context
Bump golang LRU to use newer version with Generics 

We suspect that there's an unsafe conversion between the byte key in the
map code. With the newer version, it supports generics so we need to
define the types for the underlying LRU instead of interface (where we
suspect the issue to be)

## Testing performed to validate your change

Ran two clusters, one 35 node one and a smaller one. Saw app hash issues
but none were related to the LRU issue we saw and the app hash was
repro'd in a cluster not running this new version so it's unlikely to be
releated.


![image](https://user-images.githubusercontent.com/18161326/216687552-417e7b2c-ee51-4835-a460-7e3112d30c74.png)


![image](https://user-images.githubusercontent.com/18161326/216687621-9b9f94b8-77a6-4bba-aa2a-ee5385329f46.png)