Update Misleading Code Naming (#295)
## Describe your changes and provide context

```
The pointed locations provide a misleading naming for code elements. In particular:

The Sei blockchain supports transactions signed according to the SR25519 scheme. However, the code at the pointed location accounts, for the verification cost, the constant parameter meant for signatures based on the ED25519 scheme. Since costs between the two schemes could actually be different, such constant may be misleading when configuring the parameter as it is meant for a different algorithm. Moreover it can create confusion if ED25519 needs to be supported in the future.

The handleQueryApp function handles queries to the /app endpoint and expects the second element of the path to be one among "simulate", "version" or "snapshot". However the message included in the error in the case the path does not hit one of the allowed keys only cites "simulate" and "version" as expected values, omitting the "snapshot" one, which is also supported.

The getter1 interface is used as a handy tool to call the GetAmount method on the sdk.Msg interface by casting its type. However the getter1 name is not meaningful in respect to its usage.
```
## Testing performed to validate your change

Unit tests