Fix seidb abci query race condition (#441)
## Describe your changes and provide context
This PR fix the race condition reported by ottersec:

> The issue comes from the fact that there's a small window between
deciding to take the else case and fetch tree from scStore
[here](https://github.com/sei-protocol/sei-cosmos/blob/5a05027506dcc28cacf85e8e4239a7cc3f994c8e/storev2/rootmulti/store.go#L512)
and actually fetching it
[here](https://github.com/sei-protocol/sei-cosmos/blob/5a05027506dcc28cacf85e8e4239a7cc3f994c8e/storev2/rootmulti/store.go#L514).
Another similar race is between fetching the tree
[here](https://github.com/sei-protocol/sei-cosmos/blob/5a05027506dcc28cacf85e8e4239a7cc3f994c8e/storev2/rootmulti/store.go#L514)
and actually querying it
[here](https://github.com/sei-protocol/sei-cosmos/blob/5a05027506dcc28cacf85e8e4239a7cc3f994c8e/storev2/rootmulti/store.go#L521).
In both cases, there is a chance that Commit happens in between and
changes the content of the tree fetched. Fortunately, on querying an
incorrect tree, the check
[here](https://github.com/sei-protocol/sei-cosmos/blob/5a05027506dcc28cacf85e8e4239a7cc3f994c8e/storev2/commitment/store.go#L137)
will block the query and return error instead of an incorrect result,
thus limiting the impact of the race. While the race is extremely hard
to trigger due to the tight window (we had to add some sleep into the
code to be able to hit the bug), we're once again recommending the team
to consider the approaches below to resolve this issue once and for all.



## Testing performed to validate your change