Fixed missing handling of some system calls when running using the parallel execution engine
Some transactions are executed in parallel but not for an EVM transaction. In those cases, it can happen that system calls are performed by the WASM land. We must porperly handle those cases too.

To handle those cases, we had to modify how the re-ordering of the ordinal is performed. With new interweave system calls relative to normal transactions, we needed to make the re-ordering "inline" as transaction and/or system calls are added to the block. By doing it like this, the global tracer is always aware of the currently populated ordinal. So changes happening after all other transaction will have the correct ordinal.