Genesis export stream (#519)
## Describe your changes and provide context
Writes genesis export to a file but with each module being appended to
the file individually. We need this because we are running in an issue
where the app state is too large to hold in memory causing an OOM. We
currently assume that each individual module's app state is small enough
to hold in memory.

## Testing performed to validate your change
manual testing locally.