Fix quadratic balance query (#573)
## Describe your changes and provide context
Previously `GetAllBalances` would perform slice copy operation for each
iterator `Next` call due to how `Coin::Add` is implemented. The reason
for the copy is because `Coin:Add` needs to sort the resulting slice.
However this is unnecessary because at the end of `GetAllBalances` the
slice would sorted again anyway. This PR changes the iterator operation
to be a simple append, so that the total time complexity would be O(n)
instead of O(n^2)

## Testing performed to validate your change
unit test