Do not resolve `latest` upon error (#2414)
## Describe your changes and provide context

Currently we are defaulting to `latest` if an error happens, by doing
this we are implicitly introducing side effects into user's query by
returning `latest` instead of queried on error, a much better approach
is to fail the request.

As this change introduces an API break to `RPCContextProvider` I've done
an analysis of its usage across the codebase - it's only used in RPC
routes (both `http` and `ws`) and in instantiations of EVM servers (both
`http` and `ws`) therefore the newly introduced panic can either fail a
request with
```
{
    "jsonrpc": "2.0",
    "id": 1,
    "error": {
        "code": -32603,
        "message": "method handler crashed"
    }
}
```
or panic during server instantiation which will lead to failure to boot.

## Testing performed to validate your change

Tested locally.