Delete existing zeroed out EVM contract state (#2449)
## Describe your changes and provide context

This PR will do a gradual deletion of existing zeroed out EVM contract
state. In EndBlock, we will iterate over 100 kv pairs in the EVM
contract state prefix each block and if a zero is found the kv pair is
deleted. We will keep a checkpoint for where to pick up on the next
block. This checkpoint is saved in state so we all validators can do
this deletion in lock-step and we do not app hash.

## Testing performed to validate your change
local chain testing + unit tests + performance testing on RPC node on
pacific-1 (check that a node can iterate over 100 kv pairs per block in
<1ms without significant perf degradation).