Remove sqlite and make latest version update atomic in SS (#2468)
## Describe your changes and provide context
This PR will remove SQLITE as a backend for SS store since it's not
gonna be used any more.
At the same time, it fixes some edge cases where SS store version bump
might happen when not all changes are flushed for that given version.
With this change, a single batch writes include both the version bump as
well as all changes for all modules at a given block height.
## Testing performed to validate your change
Added unit test