reverted semantics of ParseAddressString (#2470)
https://github.com/sei-protocol/sei-tendermint/pull/308 has accidentally
changed the semantics of ParseAddressString from resolving the tcp
address to parsing tcp address. As a side effect, support for DNS names
in the config has been removed.

This PR reverts to the original semantics to support DNS names again.
Also renamed the functions to better describe what they do.