Integrate watermark into evmrpc (#2465)
## Describe your changes and provide context

A common workflow that users do is to query the latest height and then
query information about that height such as events/logs. However, since
we have an option to do async writes for receipts, the events/logs may
not be available yet for that height, causing the query to fail.

In general, we want all of our heights outputted by our RPC endpoints to
be aligned with each other. We have introduced the concept of a
watermark (https://github.com/sei-protocol/sei-db/pull/116) across
Tendermint block store, SS Store, and receipt store. Pulling this
watermark across all 3 dbs should allow the evmrpc to output the minimum
height available across these dbs for a latest height. This will ensure
that all data for a given height outputted by our RPCs is available
across all our endpoints.

This PR introduces a WatermarkManager that manages the heights of these
DBs, which the evmrpc endpoints can use to get the latest heights that
are ready.

## Testing performed to validate your change
unit tests + manual testing