Fix gap nonce inclusion (#2491)
## Describe your changes and provide context

### Intro
Mempool is split into two parts - pending and active.

Every transaction gets classified (in the ante handler) as either
pending or active.
Transaction gets classified as pending if it its nonce is a skip nonce.
Transaction gets classified as active if its nonce is the next nonce.

Classification is done based on latest nonce from state.
Transaction can get promoted from pending to active only upon block
commitment via
[`mempool.Update`](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L634).

Promotion is done based on latest nonce from state + pending.
Transaction becomes eligible for promotion by way of another transaction
filling the nonce gap via updating the pending nonces.
Every transactions’ nonce gets added to pending nonces in
[`mempool.CheckTx`](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L303).
Pending nonces are not the same as the pending part of the mempool -
every transaction has its nonce included into pending nonces but it does
not have to be a pending transaction.

Inclusion to active/pending mempool happens some time after
transaction’s nonce is added to pending nonces (it does happen in the
same call path).

Because of this we’ve got a rollback mechanism that removes the
transaction’s nonce from pending nonces if transaction’s inclusion to
active/pending mempool fails among other things.

RC of the issue is that the rollback mechanism is not invoked in every
failure scenario which means that transaction's nonce can continue to
reside in pending nonces even if the transaction itself got rejected
from the mempool.

Examples of failure scenarios that do not invoke the rollback:

[[1.](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L439)]
[[2.](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L767)]
[[3.](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L818)]
[[4.](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L821)]

### Issue Scenario

1. Evaluation of `N+1` starts in the mempool -> its nonce is added to
pending nonces -> it lands into the pending pool.
2. Evaluation of `N` starts in the mempool -> its nonce is added to
pending nonces (this makes `N+1` eligible for promotion to active
mempool next time
[`mempool.Update`](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L634)
is called, for
[`mempool.Update`](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L634)
to be called evaluation of `N` needs to finish) -> it fails to get
included into the active mempool and the scenario does not invoke the
pending nonce rollback mechanism.
3.
[`mempool.Update`](https://github.com/sei-protocol/sei-chain/blob/94f51a514582889c8af929698850d0032d3e74c1/sei-tendermint/internal/mempool/mempool.go#L634)
runs and promotes `N+1` to active mempool which makes it eligible to be
included in a block.
4. `N+1` gets included in a block without `N` being included prior.
### Solution
Add the pending nonce only after transaction gets included successfully
into pending/active mempool.

## Testing performed to validate your change

I've created a local test harness that reproduced the gap nonce issue on
main - it is unable to reproduce the issue on this branch.