Remove ABCI socket/grpc functionality (#2534)
## Describe your changes and provide context

###  Intro

This PR is the first in the series of addressing our domain type state.
We will be working on a type by type starting out with the transaction.
The first issue we want to resolve is passing of domain types throughout
our codebase as encoded byte arrays.

### Background

Primary reason why we're passing structures encoded as byte arrays
internally is the fact that we've got an artificial boundary - ABCI - it
became an artificial boundary since we switched to monorepo and lost the
need for cross version compatibility between modules on boundaries of
ABCI.

ABCI boundaries are using proto types mainly because of backward/forward
compatibility.
Since we lost the need for cross version compatibility on the boundaries
of ABCI we are only left with cons of proto therefore the first thing we
need to do is migrate from proto to analogue Go types on ABCI
boundaries.

### Context

This PR allows us to wipe entire proto structures by removing a
deprecated way of ABCI intermodule communication via server and client,
it is adjacent to our typification push and allows for a really pretty
side effect.

Components that are removed:
- ABCI server
- ABCI cli (`abci-cli`)
- ABCI grpc/socket client

### Next steps

1. Move from proto types to analogue Go types on ABCI boundaries - this
should be done everywhere but first I'll be working on types that embed
a transaction as an byte array:
-
[`RequestCheckTx`](https://github.com/sei-protocol/sei-chain/blob/a38252bda9dc1e27391e99c8bba37b10e4da2813/sei-tendermint/proto/tendermint/abci/types.proto#L84)
-
[`RequestDeliverTx`](https://github.com/sei-protocol/sei-chain/blob/a38252bda9dc1e27391e99c8bba37b10e4da2813/sei-tendermint/proto/tendermint/abci/types.proto#L204)
-
[`RequestGetTxPriorityHint`](https://github.com/sei-protocol/sei-chain/blob/a38252bda9dc1e27391e99c8bba37b10e4da2813/sei-tendermint/proto/tendermint/abci/types.proto#L216)
-
[`TxResult`](https://github.com/sei-protocol/sei-chain/blob/a38252bda9dc1e27391e99c8bba37b10e4da2813/sei-tendermint/proto/tendermint/abci/types.proto#L468)
-
[`TxRecord`](https://github.com/sei-protocol/sei-chain/blob/a38252bda9dc1e27391e99c8bba37b10e4da2813/sei-tendermint/proto/tendermint/abci/types.proto#L475)
2. Instead of passing transaction as a byte array start passing it as
the most appropriate currently existent type e.g.
[`sdk.Tx`](https://github.com/sei-protocol/sei-chain/blob/06a4e242bf80fff303be607734e121bd2f0f6916/sei-cosmos/types/tx_msg.go#L39).
3. Redefine the domain type either by converging existent use cases or
by splitting the Cosmos and Ethereum use cases or both.

## Testing performed to validate your change

CI/CD.