PeerManager rewrite (#2539)
Rewrite of the PeerManager logic to make it simple, predictable and
race-condition resistant. There are further improvements to be done, but
need to keep compatibility with the current p2p implementation. With
this pr:
* the scoring system will be scraped entirely, in favor of separately
managing persistent/unconditional/blocksync peers (trusted), and peers
learned from the network (not trusted)
* peerstore will change its purpose: it used to be used to just store
known peer addresses. Now it will be used to store addresses we
successfully connected to. The only goal of peerstore now will be to
make a node able to reestablish its recent connections on restart
* node will only broadcast addresses of nodes it has successfully
connected to (so that it does not broadcast unverified data any more)
* the number of addresses per peer that peermanager keeps will be
bounded (it is unbounded currently)
* PeerManager will be an implementation detail of the Router.

---------

Co-authored-by: Jeremy Wei <jeremy.t.wei@gmail.com>