Refactor DeliverTx ante (#2569)
## Describe your changes and provide context
Refactor the ante handler logics for DeliverTx into distinct flows. No
check is supposed to be added or removed in this PR, but the ordering of
checks have changed, which would result in different error codes for
transactions that are invalid in multiple ways. As a result, this change
is consensus breaking.

The next step is to simplify the actual execution path of DeliverTx

## Testing performed to validate your change
refactor; existing tests