consensus WAL rewrite (#2582)
New implementation uses the same representation as the old one, so it is
backward compatible.
* Made repair an implementation detail of the WAL.
* Made WAL guarantee crash-tolerant atomicity (under assumption that the
file system guarantees crash-tolerant atomic appends; if not, torn
append writes are detected best effort by using crc32)
* Added file locking to prevent concurrent access to WAL
* Removed the background fsync task - background syncing is executed by
the file system anyway (afaik), and we fsync multiple times a second
(before signing) anyway.
* Removed autofile and its weird semantics of autoclosing files which
are not accessed after a delay - afaiu the idea was that one could hold
ephemeral descriptors to all WAL files, not just the head, at all times.
I've replaced it with explicit fd management.
* simplified consensus WAL reads to only support reading last height
messages - it is the only case used anyway.