Always set gas meter for every transaction (#2583)
## Describe your changes and provide context
Setting gas meter should be the very first thing when handling a
transaction. If that's not the case and the handler exits for whatever
reason, the upstream logic will use the common gas meter set at the
beginning of `FinalizeBlock` as the transaction's gas meter, which can
cause unexpected results.

## Testing performed to validate your change
local cluster