refactor: mempool: remove unused totalCheckTxCount (#2589)
## Describe your changes and provide context

Drops the unused `totalCheckTxCount` counter from the mempool. This
metric tracks the total number of transactions that the mempool asks the
app to check.

The figure isn't ever read, and I can't really think of a reason why we
might ever want to use it, so I propose dropping it.

## Testing performed to validate your change

`make test` succeeds.