[CON-143] fix: deflake TestRouter_dialPeer_Reject (#2603)
## Describe your changes and provide context

This test can flakily fail with the following error:

```
--- FAIL: TestRouter_dialPeer_Reject (0.14s)
    --- FAIL: TestRouter_dialPeer_Reject/incompatible_channels (0.02s)
        router_test.go:548: want EOF, got mconn.Run: recvRoutine: protoReader.ReadMsg(): read tcp 127.0.0.1:41769->127.0.0.1:44942: read: connection reset by peer
```

The issue is that the test only wants an EOF error to signal a
connection being closed, but you can probably get a bunch of different
errors. This PR weakens the fail condition to `utils.IgnoreCancel(err)
== nil`, matching the pattern we already have in TestRouter_AcceptPeers
(see
[here](https://github.com/sei-protocol/sei-chain/blob/4bc58d5aee5cdf4ed02ed9228138b7960e82bc69/sei-tendermint/internal/p2p/router_test.go#L383)).

This PR also makes the same change to `TestRouter_EvictPeers`, which
might flake in a similar way.

## Testing performed to validate your change

`make test` passes