[giga] add executor interfaces for VM (#2601)
## Describe your changes and provide context
EVM implementations roughly adopt the following paradigm:
```
some preprocessing
calls the VM
- the VM starts doing stuff internally
some postprocessing
```
This PR defines the interfaces for the `calls the VM` part. The goal is
so that we can relatively easily swap out the VM without changes to the
rest of the codebase.
It also defines the storage interface for preprocessing/postprocessing
to use without worrying about the underlying storage implementation
(which is VM-specific).

The PR also provides implementation of the interfaces by Geth and EVMC,
but not storage interface implementation yet.

## Testing performed to validate your change
no functional changes.